refactor: merge 'github' package inside the 'web' app
This independent package is not useful as such, as it is only used by the app. Removing the package to simplify the project structure.
