+++ apps/web/.eslintrc.cjs
@@ -15,4 +15,10 @@ module.exports = {
       "./tsconfig.cloudflare.json",
     ],
   },
+  rules: {
+    "@typescript-eslint/no-unused-vars": [
+      "error",
+      { args: "none" },
+    ]
+  },
 };
