+++ apps/web/package.json
@@ -9,7 +9,7 @@
     "preview": "vite preview",
     "typecheck": "tsc -b",
     "lint": "eslint \"src/**/*.{ts,tsx}\"",
-    "codegen": "react-router typegen && wrangler types",
+    "codegen": "react-router typegen && wrangler types && graphql-codegen --config ./codegen.ts",
     "test": "vitest run",
     "coverage": "vitest run --coverage"
   },
@@ -17,8 +17,9 @@
     "@blueprintjs/core": "^5.19.0",
     "@blueprintjs/icons": "^5.22.0",
     "@blueprintjs/select": "^5.3.20",
+    "@graphql-typed-document-node/core": "^3.2.0",
+    "@octokit/plugin-throttling": "^11.0.1",
     "@react-router/node": "^7.6.0",
-    "@repo/github": "workspace:*",
     "@tanstack/react-query": "^5.75.4",
     "@tanstack/react-query-devtools": "^5.75.4",
     "@tanstack/react-query-persist-client": "^5.75.4",
@@ -26,11 +27,13 @@
     "clsx": "^2.1.1",
     "dexie": "^4.0.9",
     "dexie-react-hooks": "^1.1.7",
+    "graphql": "^16.11.0",
     "idb-keyval": "^6.2.1",
     "isbot": "^5.1.28",
     "javascript-time-ago": "^2.5.10",
     "localforage": "^1.10.0",
     "normalize.css": "^8.0.1",
+    "octokit": "^5.0.2",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "react-router": "^7.6.0",
@@ -41,6 +44,10 @@
   },
   "devDependencies": {
     "@codecov/vite-plugin": "^1.9.0",
+    "@graphql-codegen/cli": "^5.0.5",
+    "@pollyjs/adapter-fetch": "^6.0.6",
+    "@pollyjs/core": "^6.0.6",
+    "@pollyjs/persister-fs": "^6.0.6",
     "@react-router/dev": "^7.6.0",
     "@repo/eslint-config": "workspace:*",
     "@repo/typescript-config": "workspace:*",
