+++ apps/web/src/components/PullRow.tsx
@@ -1,6 +1,6 @@
 import { Tooltip, Tag, Icon } from "@blueprintjs/core";
 import TimeAgo from "./TimeAgo";
-import type { Pull } from "@repo/github";
+import type { Pull } from "../lib/github/types";
 import IconWithTooltip from "./IconWithTooltip";
 import { computeSize } from "../lib/size";
 import styles from "./PullRow.module.scss";
