+++ apps/web/src/github.tsx
@@ -1,4 +1,4 @@
-import { type GitHubClient, DefaultGitHubClient } from "@repo/github";
+import { type GitHubClient, DefaultGitHubClient } from "./lib/github/client";
 
 // Expose a GitHub client as a singleton.
 export const gitHubClient: GitHubClient = new DefaultGitHubClient();
