--- packages/github/README.md
@@ -1,3 +1,0 @@
-# packages/github
-
-This package contains the GitHub client.
