--- packages/github/package.json
@@ -1,38 +1,0 @@
-{
-  "name": "@repo/github",
-  "version": "0.0.0",
-  "type": "module",
-  "license": "MIT",
-  "main": "src/index.ts",
-  "files": [
-    "src",
-    "generated"
-  ],
-  "scripts": {
-    "typecheck": "tsc -b",
-    "lint": "eslint \"**/*.ts\"",
-    "test": "vitest run",
-    "coverage": "vitest run --coverage",
-    "codegen": "graphql-codegen --config ./codegen.ts"
-  },
-  "devDependencies": {
-    "@graphql-codegen/cli": "^5.0.5",
-    "@pollyjs/adapter-fetch": "^6.0.6",
-    "@pollyjs/core": "^6.0.6",
-    "@pollyjs/persister-fs": "^6.0.6",
-    "@repo/eslint-config": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
-    "@types/node": "^22.5.0",
-    "@vitest/coverage-istanbul": "^2.0.5",
-    "eslint": "^8.57.0",
-    "typescript": "^5.5.4",
-    "vitest": "^2.0.5"
-  },
-  "dependencies": {
-    "@graphql-typed-document-node/core": "^3.2.0",
-    "@octokit/plugin-throttling": "^9.3.1",
-    "graphql": "^16.11.0",
-    "octokit": "^4.0.2",
-    "remeda": "^2.0.5"
-  }
-}
