--- packages/github/src/index.ts
@@ -1,4 +1,0 @@
-export * from "./attention";
-export * from "./client";
-export * from "./search";
-export * from "./types";
