--- packages/github/tsconfig.build.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@repo/typescript-config/base.json",
-  "include": ["src", "tests", "generated/**/*"],
-  "compilerOptions": {
-    "composite": true,
-    "target": "ES2022",
-    "module": "ES2022",
-    "moduleResolution": "bundler",
-    "rootDirs": ["./src", "./generated"],
-    "noEmit": true,
-  },
-}
