--- packages/github/tsconfig.node.json
@@ -1,13 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/tsconfig",
-  "extends": "@repo/typescript-config/base.json",
-  "include": [".eslintrs.cjs", "codegen.ts", "vitest.config.ts"],
-  "compilerOptions": {
-    "composite": true,
-    "types": ["node"],
-    "target": "ES2022",
-    "module": "ES2022",
-    "moduleResolution": "bundler",
-    "noEmit": true,
-  }
-}
