--- packages/github/vitest.config.ts
@@ -1,11 +1,0 @@
-import { defineConfig } from "vitest/config";
-
-export default defineConfig({
-  test: {
-    include: ["tests/**/*.test.ts"],
-    coverage: {
-      include: ["src/**"],
-      provider: "istanbul",
-    },
-  },
-});
