--- packages/typescript-config/vite.json
@@ -1,18 +1,0 @@
-{
-  "extends": "./base.json",
-  "compilerOptions": {
-    "target": "es2015",
-    "lib": ["es2023", "DOM", "WebWorker"],
-    "useDefineForClassFields": true,
-    "module": "ESNext",
-    "sourceMap": true,
-    "resolveJsonModule": true,
-    "moduleResolution": "Bundler",
-    "noEmit": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true,
-    "allowImportingTsExtensions": true,
-    "jsx": "react-jsx"
-  }
-}
